writing jitted routine with new pydantic classes
